requirements.txt
langchain>=0.0.300
openai>=0.27.0
Pillow>=9.0.0
